requirements.txt
streamlit
requests
python-dotenv
tqdm
pandas
